tests/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
exceptiongroup==1.2.2
idna==3.8
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.3
tomli==2.0.1
urllib3==2.2.2
